pulled from upstream/master, reworked controller for admin
