+++ pokemon-spring-mvc/src/main/resources/Texts.properties
@@ -63,3 +63,5 @@ main.page.text=Welcome to Pokemon Leage Paricipation Manager.
 project.page=Project page.
 catch.them.all=Gotta catch 'em all!
 log.out=Log out
+trainer.setAdmin=Set as Admin
+trainer.unsetAdmin=
